+++ prql-compiler/src/ast/mod.rs
@@ -2,3 +2,6 @@
 //!
 pub mod pl;
 pub mod rq;
+mod span;
+
+pub use span::Span;
